Put configureAbly in _app.js and refactor
