--- README_old.md
@@ -1,17 +1,0 @@
-# Ably/Next Realtime News
-
-This is a demo app using [NextJS](https://nextjs.org/) and [Ably](https://ably.com) with the Ably's [react-hooks](https://github.com/ably-labs/react-hooks) `npm` module.
-
-It demonstrates use of:
-
-- Pub/sub messaging
-- Channel presence
-- History
-
-You can test it out by visiting the [Vercel deployment](https://ably-next-vercel-news.vercel.app/) in multiple pages.
-
-To run it locally, fork this repo and add an `.env.local` file containing your Ably API key:
-
-```
-ABLY_API_KEY=your.api-key:goes-here
-```
