+++ components/Presence.js
@@ -3,15 +3,13 @@ import { usePresence } from "@ably-labs/react-hooks";
 import styles from "../styles/Home.module.css";
 
 /* Retrieves the Presence set from the "news-list" channel and lists
-the members using their client Id */
-const AblyPresence = (props) => {
+the members using their randomly-assigned client Id */
+const Presence = (props) => {
   const [presenceData] = usePresence("news-list");
 
   const presenceList = presenceData.map((member, index) => {
-    /* Is this member of the Presence set the current user? If so,
-    mark it as "(me)" */
-    console.log(props.user);
-    const isItMe = member.clientId === props.user.clientId ? "(me)" : "";
+    // Is this member of the Presence set the current user?
+    const isItMe = member.clientId === props.auth.clientId ? "(me)" : "";
     return (
       <li key={index}>
         <span className={styles.participant}>{member.clientId}</span>{" "}
@@ -23,4 +21,4 @@ const AblyPresence = (props) => {
   return <div>{presenceList}</div>;
 };
 
-export default AblyPresence;
+export default Presence;
