+++ example.env
@@ -1,3 +1,5 @@
+# This will change after deployment
+NEXT_PUBLIC_HOSTNAME=http://localhost:3000
 # Server key must have Publish and History permissions - use app's Root key
 ABLY_SERVER_API_KEY=your.api-key:goes-here
 # Client key should have Subscribe and Presence permissions only
