+++ next.config.js
@@ -1,10 +1,12 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {
+
+const withTM = require("next-transpile-modules")(["@ably-labs/react-hooks"]);
+const nextConfig = withTM({
   reactStrictMode: true,
   images: {
     dangerouslyAllowSVG: true,
     domains: ["static.ably.dev"],
   },
-};
+});
 
 module.exports = nextConfig;
