+++ package.json
@@ -14,6 +14,7 @@
     "bad-words": "^3.0.4",
     "date-and-time": "^2.3.1",
     "next": "12.1.5",
+    "next-transpile-modules": "^9.0.0",
     "react": "18.0.0",
     "react-dom": "18.0.0",
     "unique-names-generator": "^4.7.1"
